requirements.txt
torch>=2.0.0
torchvision>=0.15.0
sacrebleu>=2.3.1
rouge-score>=0.1.2
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
Pillow>=9.5.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.3.0
